+++ src/hpack/hpack.c
@@ -3,6 +3,7 @@
 #include "hpack/decoder.h"
 #include "hpack/encoder.h"
 #include "hpack/tables.h"
+#include "config.h"
 #define LOG_MODULE LOG_MODULE_HPACK
 
 
@@ -19,7 +20,7 @@
  * Output:
  *      returns the amount of octets in which the pointer has move to read all the headers
  */
-int decode_header_block(hpack_states_t *states, uint8_t *header_block, uint8_t header_block_size, header_list_t *headers)//header_t* h_list, uint8_t * header_counter)
+int decode_header_block(hpack_states_t *states, uint8_t *header_block, int header_block_size, header_list_t *headers)//header_t* h_list, uint8_t * header_counter)
 {
     return hpack_decoder_decode_header_block(states, header_block, header_block_size, headers);
 }
