+++ src/hpack/hpack.h
@@ -12,7 +12,7 @@
 
 
 
-int decode_header_block(hpack_states_t *states, uint8_t *header_block, uint8_t header_block_size, header_list_t *headers);
+int decode_header_block(hpack_states_t *states, uint8_t *header_block, int header_block_size, header_list_t *headers);
 
 int encode(hpack_states_t *states, char *name_string, char *value_string,  uint8_t *encoded_buffer);
